remove nested file structure for routes, more cleanups for routes soon
