--- src/routes/index.ts
@@ -1,9 +1,0 @@
-import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
-
-export default async (app: FastifyInstance) => {
-    app.get('/', async (req: FastifyRequest, reply: FastifyReply) => {
-        return reply.send({
-            message: 'Welcome to C2C',
-        });
-    });
-};
